tests: network utils are used by integration tests only
